convert proto= option to use netids rather than a protoname
Patchwork-id: 21843
O-Subject: [RHEL6.0 PATCH] BZ#545973: NFS: convert proto= option to use netids
	rather than a protoname
Bugzilla: 545973
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Peter Staubach <staubach@redhat.com>

This patch is being pushed to mainline by Trond for 2.6.33. Since
since the goal is to support client side IPv6 in RHEL6 GA, we'll
need to take this into RHEL6's kernel soon. Tested on a RHEL6
alpha install with a 2.6.32 kernel and didn't see any regressions
for v4 mounting. The same patch has been tested on a rawhide install
with Chuck Lever's nfs-utils patches, and that works as expected.

-------------------------[snip]------------------------

Solaris uses netids as values for the proto= option, so that when
someone specifies "tcp6" they get traffic over TCP + IPv6. Until
recently, this has never really been an issue for Linux since it didn't
support NFS over IPv6. The netid and the protocol name were generally
always the same (modulo any strange configuration in /etc/netconfig).

The solaris manpage documents their proto= option as:

    proto= _netid_ | rdma

This patch is intended to bring Linux closer to how the Solaris proto=
option works, by declaring a static netid mapping in the kernel and
converting the proto= and mountproto= options to follow it and display
the proper values in /proc/mounts.

Much of this functionality will need to be provided by a userspace
mount.nfs patch. Chuck Lever has a patch to change mount.nfs in
the same way. In principle, we could do *all* of this in userspace but
that would mean that the options in /proc/mounts may not match the
options used by userspace.

The alternative to the static mapping here is to add a mechanism to
upcall to userspace for netid's. I'm not opposed to that option, but
it'll probably mean more overhead (and quite a bit more code). Rather
than shoot for that at first, I figured it was probably better to
start simply.

Comments welcome.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>