Fix potential race in glock code
Patchwork-id: 21845
O-Subject: [RHEL 6] GFS2: Fix potential race in glock code (bz #546279)
Bugzilla: 546279
RH-Acked-by: Benjamin Marzinski <bmarzins@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The following patch is in upstream (and RHEL5) and just missed the
2.6.32 deadline, so it needs to be back ported into RHEL6. It fixes
bz #546279.

I'm about to send a number of follow up patches, the application
order of the patches should be:

1. This patch
2. The GFS2 ACL/xattr patch series
3. The GFS2 quota patch series
4. Anything else

This matches the patch ordering in upstream.

Steve.

>From 7e71c55ee73988d0cb61045660b899eaac23bf8f Mon Sep 17 00:00:00 2001
From: Steven Whitehouse <swhiteho@redhat.com>
Date: Tue, 22 Sep 2009 10:56:16 +0100
Subject: GFS2: Fix potential race in glock code

We need to be careful of the ordering between clearing the
GLF_LOCK bit and scheduling the workqueue.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>