Hook gfs2_quota_sync into VFS via gfs2_quotactl_ops
Patchwork-id: 21857
O-Subject: [RHEL6] [02/13] GFS2: Hook gfs2_quota_sync into VFS via
	gfs2_quotactl_ops
Bugzilla: 298561
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The plan is to add further operations to the gfs2_quotactl_ops
in future patches. The sync operation is easy, so we start with
that one.

We plan to use the XFS quota control functions because they more
closely match the GFS2 ones.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>