Clean up gfs2_adjust_quota() and do_glock()
Patchwork-id: 21864
O-Subject: [RHEL6] [08/13] GFS2: Clean up gfs2_adjust_quota() and do_glock()
Bugzilla: 298561
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

Both of these functions contained confusing and in one case
duplicate code. This patch adds a new check in do_glock()
so that we report -ENOENT if we are asked to sync a quota
entry which doesn't exist. Due to the previous patch this is
now reported correctly to userspace.

Also there are a few new comments, and I hope that the code
is easier to understand now.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>