Fix -o meta mounts for subsequent mounts
Patchwork-id: 21884
O-Subject: [RHEL6] GFS2: Fix -o meta mounts for subsequent mounts (bz #546664)
Bugzilla: 546664
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This is a bug fix from upstream which we need to include in RHEL6.
It fixes an issue with mounts of the meta root of a gfs2 filesystem
which was failing as described below. Tested in upstream. This
fixes bz #546664

Steve.

>From f55073ff1eaf99f6b3bc62134a456638bca043a3 Mon Sep 17 00:00:00 2001
From: Steven Whitehouse <swhiteho@redhat.com>
Date: Mon, 28 Sep 2009 10:30:49 +0100
Subject: GFS2: Fix -o meta mounts for subsequent mounts (i.e. all but the first one)

We have a long term plan to use the "-o meta" flag to GFS2 mounts to
access the alternate root which is used to store metadata for a GFS2
filesystem. This will allow us to eventually remove support for the
gfs2meta filesystem type (which is in any case just a "front end" to
the gfs2 filesystem type with the meta/master root).

Currently the "-o meta" option is only taken into account on the
initial mount of the filesystem. Subsequent mounts of the same
filesystem (i.e. on the same device) result in basically the same
as bind mounting the root of the original mount.

This patch changes that by using what is more or less a copy
of get_sb_bdev() and extending it so that it will take into
account the alternate root in all cases. The main difference
is that we have to parse the mount options a bit earlier. We can
then use them to select the appropriate root towards the end of
the function.

In addition this also fixes a bug where it was possible (but certainly
not desirable) to set different ro/rw options for the meta root
when mounted via the gfs2meta fs compared with the original mount.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Cc: Alexander Viro <aviro@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>