kernel: clear high-order bits after switching to 64-bit mode
Patchwork-id: 21888
O-Subject: [RHEL6 PATCH 1/5] s390 - kernel: clear high-order bits after
	switching to 64-bit mode
Bugzilla: 546314
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
IPLing a Linux system fails when the IPL CLEAR option
is not specified. Because the kernel fails early in the
boot process, no kernel messages are printed to the
console.  In rare cases, zfcpdump might fail to start dumping.

When IPLing without the IPL CLEAR option, registers might
contain random values in the high-order bits. When the
kernel switches to 64-bit mode, the registers are not
cleared. This results in (addressing) exceptions and the
kernel might enter an interrupt loop.

To solve the problem, initialize the high-order bits of the
registers with zeroes right after the kernel switched to
64-bit mode.

Bugzilla
--------
BZ 546314
https://bugzilla.redhat.com/show_bug.cgi?id=546314

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commit;h=cf87b7439ec81b9374e7772e44e9cb2eb9e57160

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>