dasd: support DIAG access for read-only devices
Patchwork-id: 21890
O-Subject: [RHEL6 PATCH 3/5] s390 - dasd: support DIAG access for read-only
	devices
Bugzilla: 546309
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
If a DASD is attached as read-only under z/VM and it is configured
to be used with the DIAG discipline, then setting the device online
will fail.
The message log will show a message 'DIAG initialization failed (rc=4)'.

The DIAG return code 4 indicates that the initialization was successful
and the device is read-only, but the DASD device driver interprets this
as an error.

The solution is to fix the DIAG device initialization to accept a return
value of 4 as indication of success.

Bugzilla
--------
BZ 546309
https://bugzilla.redhat.com/show_bug.cgi?id=546309

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commit;h=22825ab7693fd29769518a0d25ba43c01a50092a

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>