Build in standard PCI hotplug support
Patchwork-id: 21927
O-Subject: [PATCH 1/2] RHEL 6 config: Build in standard PCI hotplug support
Bugzilla: 517050
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

There's no way to autodetect whether a system implements either native
PCIe hotplug or ACPI-mediated PCI hotplug. However, some devices require
this functionality to work correctly (notably many Expresscard bays in
laptops). The only way to obtain functionality without requiring manual
configuration is to build them in.

Customers have previously been opposed to this. One reason was that in the
past, having these drivers linked in would cause the ACPI code to bind
before native PCIe hotplug. This was fixed in the upstream kernel in order
to give native PCIe hotplug the opportunity to bind first. It will make an
_OSC call through the ACPI firmware in order to request that the firmware
relinquish control to the OS. If the firmware refuses, the driver will not
bind and control will fall back to the ACPI hotplug driver.

A followup patch will be provided in order to allow users who want to use
ACPI-mediated hotplug to do so even if their hardware and firmware support
native PCIe hotplug.

Note that this simply changes the configuration to match that shipped in
Fedora since F11.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>