dell-laptop: Update rfkill state on switch change
Patchwork-id: 21931
O-Subject: [PATCH 2/2] [RHEL6 PATCH] dell-laptop: Update rfkill state on switch
	change
Bugzilla: 547892
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

This is a backport of an upstream-destined patch that replaces the
dell-laptop patch in the F12 kernel. It's been respun so that the kABI
requirements match the direction upstream's going, rather than being
specific to RHEL.

The rfkill interface on Dells only sends a notification that the switch
has been changed via the keyboard controller. Add a filter so we can
pick these notifications up and update the rfkill state appropriately.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>