zfcp: Assign scheduled work to driver queue
Patchwork-id: 21960
O-Subject: [RHEL6 PATCH 2/6] [s390] zfcp: Assign scheduled work to driver queue
Bugzilla: 547377
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
The following kernel message is logged:
"kobject_add_internal failed for 0x5005076401a07163 with -EEXIST,
don't try to register things with the same name in the same directory."

The port_scan work was scheduled to the work_queue provided by the kernel.
This resulted on SMP systems to a likely situation that more than one
scan_work were processed concurrently.  This is not required and opens
the possibility of race conditions between the removal of invalid ports
and the enqueue of just scanned ports.

The solution is to synchronize the scan_work tasks by scheduling them
to adapter local work_queue.

Bugzilla
--------
BZ 547377
https://bugzilla.redhat.com/show_bug.cgi?id=547377

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=9eae07ef6bb5988163d8bb82cd952905db47b721

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>