zfcp: fix ELS ADISC handling to prevent QDIO errors
Patchwork-id: 21959
O-Subject: [RHEL6 PATCH 3/6] [s390] zfcp: fix ELS ADISC handling to prevent QDIO
	errors
Bugzilla: 547385
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
When issuing a ELS ADISC to test the connection to a remote port,
this could lead to a qdio error and adapter recovery.
The ELS ADISC data does not satisfy the hardware alignment
requirements, leading to the error and recovery.

Introduce a kmem_cache for the ELS ADISC request data and use
zfcp_cache_hw_align to align the data like other requests sent
to the hardware.

Bugzilla
--------
BZ 547385
https://bugzilla.redhat.com/show_bug.cgi?id=547385

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=ee744622c65cd66824e8dd1b9509e515c800de14

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>