dasd: remove strings from s390dbf
Patchwork-id: 21966
O-Subject: [RHEL6 PATCH 2/3] [s390] dasd: remove strings from s390dbf
Bugzilla: 547735
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
With some messages in the s390dbf an kernel oops can occur while
reading messages from the debug feature.
Strings in the s390dbf are only copied as links. So the bus id
for example has only the lifetime of the device it was generated on.

Do not use strings as identifier of DASD devices. Instead, use a
numeric representation of the bus id.

Bugzilla
--------
BZ 547735
https://bugzilla.redhat.com/show_bug.cgi?id=547735

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=b8ed5dd54895647c2690575aad6f07748c2c618a

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>