dasd: let device initialization wait for LCU setup
Patchwork-id: 21965
O-Subject: [RHEL6 PATCH 3/3] [s390] dasd: let device initialization wait for LCU
	setup
Bugzilla: 547735
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
The first DASD that is set online for a specific logical control
unit has to do certain setup steps on the storage server to make
full use of it, for example, it will enable PAV.
The features and characteristics reported by the storage server
will depend on this setup.

When other devices on the same storage server are set online before
the setup of the first DASD is completed some features might be missing.

Solution: Let all devices on an LCU wait for the setup of first device.

Bugzilla
--------
BZ 547735
https://bugzilla.redhat.com/show_bug.cgi?id=547735

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=f4ac1d0255748fe0f8e128a26b1c29490cae5c08

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>