kernel: enable hibernation support on s390x
Patchwork-id: 21976
O-Subject: [RHEL6 PATCH 1/4] [s390] kernel: enable hibernation support
Bugzilla: 547390
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
This commit turns on CONFIG_HIBERNATION for the default s390x kernel.

Notes
-----
This bugzilla is contrary to BZ 521135 which is a specific request
to disable CONFIG_HIBERNATION.

Turning on CONFIG_HIBERNATION conflicts with the CONFIG_MEMORY_HOTPLUG
kernel config option.  With 2.6.32 both options can be switched on,
the respective commit ID is:
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=6ad696d2cf535772dff659298ec7e7260e344595

For s390, the user can use either hibernation or memory hotplug
but not both.  A check ensures that hibernation requests are
refused when memory has been added or removed.

The change also includes Kconfig changes for enabling kernel options
for both memory hotplug and hibernation.

Please see the commit link below:
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=039979049834bde56f67f8078c802b416bd4763c
(This commit is in the upstream kernel as of version 2.6.31
and the Kconfig changes have been modified by commit
6ad696d2cf535772dff659298ec7e7260e344595).

Bugzilla
--------
BZ 547390
https://bugzilla.redhat.com/show_bug.cgi?id=547390

Related bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=516805

Please ACK.

With best regards,

        Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>